gweb: Fix crash on malformed http response
If an http response has a line starting with whitespace before the first
valid header field, g_hash_table_lookup() is called with a NULL key,
which is invalid.  Fix this by checking for a valid last_key before
attempting to process continuation lines.

Based on a patch by Tomasz Bursztyka for an issue originally reported by
Marcel Mulder.