session: Add AllowedInterface config parameter
In order to be able to bind a session to a specific interface,
AllowedInterface config option has been added. It allows to
bind a session to a service with a specified network interface
name (e.g. eth0).

AllowedInterface is supported through both dbus and session policy.
Policy always overrides dbus.