technology: Deny P2P finding if technology is disabled
This check is added because P2P and WiFi scan end up at the same device and at
that level there is not differentiation between WiFi and P2P, they both are WiFi
devices. Therefore, when device's power is checked in device.c:device_scan(),
it is actually checking only if WiFi technology is enabled thus P2P scan will go
ahead even if P2P technology is not enabled.