firewall-nftables: Fix build with libnftnl-1.0.7
We need these updates to accommodate the changes caused by the following
commit in libnftnl-1.0.7

commit 907a9f8e5a93f5bcd449643eb3916a656d634758
Author: Pablo Neira Ayuso <pablo@netfilter.org>
Date:   Tue Dec 20 13:47:11 2016 +0100

src: get rid of aliases and compat

This machinery was introduced to avoid sudden compilation breakage of
old nftables releases. With the upcoming release of 0.7 (and 0.6 which
is now 6 months old) this is not required anymore.

Moreover, users gain nothing from older releases since they are
half-boiled and buggy.

So let's get rid of aliases now. Bump LIBVERSION and update map file.