service: Fix loose mode routing not enabled with EnableOnlineCheck option
When EnableOnlineCheck option is disabled, the rp_filter is not
updated. If there are more than one service available, the result is
that some packets are blocked.

The initial support for loose mode routing was added by
cb3e78500a25 ("service: Activate loose mode routing"). Commit
4de35cde5a93 ("service: Add EnableOnlineCheck config option") moved
the rp_filter under the EnableOnlineCheck condition.