config: Support mDNS in service configuration files
Use key "mDNS" with boolean values. True means that mDNS is enabled.
mDNS domains can be resolved and hostname is registered. False means
that all mDNS functionality for this service is disabled. Note that
not all DNS backends support mDNS: currently systemd-resolved is the
only DNS backend that works with mDNS.