scripts: Don't forceably overwrite /etc/resolv.conf
Using L+ in the tmpfiles.d snippet is overly aggressive.  These snippets
are evaluated on every boot and may be evaluated on package upgrade.
It's fine to check that the file /etc/resolv.conf exists, but forceably
overwriting an existing file or link makes it difficult for connman to
co-exist with other potential managers of /etc/resolv.conf.