config: Reduce noisy error when we try to provision hidden service from config file
Before this change upon every scan of the hidden service we got
"Network SSID not set" error in logs. This is because hidden services
have no WiFi.SSID set.

Changed function is try_provision_service from config file and of course
we can not do this for hidden services without SSID so we can report
ENOENT for them.