ntp: Report error to upper layer
Inform the timeserver layer if there was an error. Without the error
reports the timeserver code assumes all is fine.

For example if the system does not suppoert IPv6 but we get an IPv6
NTP server address, start_ntp() will fail to create an IPv6 socket.

Let's report all error back so that timeserver is able to react on the
error.

Reported-by: Craig McQueen <craig.mcqueen@innerrange.com>