vpn: Add remove function callback to VPN driver
This commit adds remove function callback to the VPN driver structure.
The prototype is int (remove*) (struct vpn_provider *provider) and is to
be registered within the VPN plugin.

This function is called from vpn/plugins/vpn.c when removing a provider
after it has been stopped. It is meant for VPNs that need to cleanup
some VPN specific content when removed or to notify other services after
removal.