vpn: Use Connect2 method only when D-Bus sender is set
This fix addresses the issue of not being able to connect a VPN if it
has autoconnect set and connection is triggered by connmand. In such
case dbus_sender is not set and call to Connect2 method of
net.connman.vpn.Connection cannot be done.

Also, print correct D-Bus method called in case of error. If the
dbus_sender is not set or is empty Connect() will be used instead of
Connect2().

This amends changes of commit 3997c1595fe44d7c85215fd1cd89ecf778c62206.