service: Re-add run_vpn_auto_connect() with increasing delays to main loop
This commit enhances run_vpn_auto_connect() to re-add itself to main
loop when there are VPNs to auto connect but VPN connection has not
completed yet. The delay will increase at each attempt, which at start
is 0 and is passed along as the user data. The delay will increase after
VPN_AUTOCONNECT_TIMEOUT_STEP attempts (30) with 1s, starting with
default value VPN_AUTOCONNECT_TIMEOUT_DEFAULT (1) that is being set
after the initial call. The value is increased until the threshold
(VPN_AUTOCONNECT_TIMEOUT_ATTEMPTS_TRESHOLD, 270) is reached. Delays
longer than 10s are not imposed.

This improves the autoconnection in situations when the network
conditions change rapidly, networks are slow to get online status, there
are lots of repetitive switches between networks and different network
types. If there is no connected default service or the default service
is a connected VPN or there are no VPNs with autoconnect set function is
removed from mainloop.

Connecting VPNs are regarded as autoconnectable to ensure re-connection
of the VPN if the network breaks during connection establishment and new
network is being connected to.