vpn-provider: Add support for plugin specific data
Support plugin specific data in provider. Setting internal data of a
plugin to provider is useful in case the VPN plugin needs to perform any
action based on the data when notify is called. For example, failure
counters can be incremented based on the sent notify information, which
may be later used for sending additional information to VPN provider.