service: Clear VPN error and state if AutoConnect is set on
Changing AutoConnect value to "true" must have the same effect as user
calling "Connect" method via D-Bus. The failure state is ignored in
do_auto_connect() if user called the VPN to connect.

This fixes the inconsistency with AutoConnect value setting, if user
sets AutoConnect to "true" clear error and set the service state to
idle, if VPN is in failure state. This makes it possible to get a
VPN connecting immediately after toggling the AutoConnect value.