vpn: Call disconnect() also for non TUN VPNs
Add the missing disconnect() call. In case where the D-Bus
'disconnect' method call is not used to shutdown the connection, we
leak the connection.

For example when the users edits the config file, we end up in the
stop_vpn() function with an active WireGuard connection.

Fixes 12f095a32eff ("vpn: Introduce VPN_FLAG_NO_DAEMON")
Reported by Christian Hewitt