vpn: Move vpn_provider_get_ident() declaration to vpn-provider.h
All the functions the VPN plugins would need should be declared in the
relevant headers. VPNs do not need to include vpn/vpn.h as the functions
declared there are not for plugins to use.

Drop also the "../vpn.h" include from OpenVPN plugin, which was the only
plugin using vpn_provider_get_ident().