vpn: Secure a race condition with flag
ConnMan crashes happens in cancel_host_resolv in plugins/vpn.c. It
seems that the function gets called twice on the same data.

cancel_host_resolv is called from 2 places - from remove_result and from
connection_destroy. remove_result is called from resolv_result
indirectly - through g_timeout_add_seconds(0,...). In connection_destroy
it is called directly as part of the cleanup routine.

I managed to trigger a situation where remove_result was called and the
resolv structure was cleaned and freed, and after that (due to me
killing connman-vpnd in an unfortunate moment), connection_destroy
kicked in and attempted to access data in resolv (NULL) pointer.

To fix the issue, I add a user_data structure element representing the
event ID of the scheduled resolv structure removal function. The
cancel_host_resolv function would then check if the resolv removal
function was scheduled, and would remove it from the loop, so it does
not get executed.