Revert "gdhcp: Make DHCP client timeouts suspend aware"
This reverts commit e17f601cf4af77bcc0b3cfe576c5d0414511e5aa.

CLOCK_BOOTTIME_ALARM wakes the device up in suspend mode. While this
fixes the problem that all DHCP lease timeouts are handled correctly
even in suspend state, it wakes up the device unnecessary. The DHCP
code needs to be made aware of the power states and act accordingly.

Because this change introduced a noticable regression, let's revert
it for the time being.